Converted main landing page to server side rendering
